feat(graph): support trackpad pinch zoom
Replaces the React onWheel handler (passive by default, so
preventDefault is a no-op) with a native addEventListener that uses
{ passive: false }. Safari and Chrome synthesize wheel events with
ctrlKey=true for two-finger pinch on Mac trackpads — without
preventDefault those gestures zoom the whole page instead of the
graph. Also normalizes deltaMode=1 (lines) to pixels so mouse wheel
and pinch deltas feel similar.
